rework the p2v classify and benchmark scripts
